# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lugendo — Travel agency management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agency-inquiries:
    post:
      operationId: createAgencyInquiry
      tags: [agency-inquiries]
      summary: Send an inquiry to an agency (requires a session)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgencyInquiryInput"
      responses:
        "201":
          description: Inquiry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgencyInquiry"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Agency not found
    get:
      operationId: listAgencyInquiries
      tags: [agency-inquiries]
      summary: List inquiries received by the caller's agency (Admin/Manager)
      responses:
        "200":
          description: Inquiries for the caller's agency
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgencyInquiryReceived"
...
  /agency-inquiries/{inquiryId}/read:
    patch:
      operationId: markAgencyInquiryRead
      tags: [agency-inquiries]
      summary: Mark an inquiry as read (Admin/Manager, own agency only)
      parameters:
        - name: inquiryId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgencyInquiry"
        "404":
          description: Not found

  # ─── AUTH ───────────────────────────────────────────────────────────────────
...
  /hotels:
    get:
      operationId: listHotels
      tags: [hotels]
      summary: List hotels in the agency catalog
      responses:
        "200":
          description: List of hotels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hotel"
    post:
      operationId: createHotel
      tags: [hotels]
      summary: Add hotel to agency catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HotelInput"
      responses:
        "201":
          description: Hotel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
  /hotels/{hotelId}:
    get:
      operationId: getHotel
      tags: [hotels]
      summary: Get hotel detail
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hotel detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"
        "404":
          description: Not found
    patch:
      operationId: updateHotel
      tags: [hotels]
      summary: Update hotel
      parameters:
        - name: hotelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HotelUpdate"
      responses:
        "200":
          description: Updated hotel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hotel"

  # ─── ACTIVITIES ─────────────────────────────────────────────────────────────
...
  /checklist-templates:
    get:
      operationId: listChecklistTemplates
      tags: [checklist-templates]
      summary: List the agency's checklist templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistTemplate"
    post:
      operationId: createChecklistTemplate
      tags: [checklist-templates]
      summary: Create a checklist template for the agency
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistTemplateInput"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistTemplate"
  /checklist-templates/{templateId}:
    patch:
      operationId: updateChecklistTemplate
      tags: [checklist-templates]
      summary: Update a checklist template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChecklistTemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChecklistTemplate"
    delete:
      operationId: deleteChecklistTemplate
      tags: [checklist-templates]
      summary: Delete a checklist template
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    AgencyInquiryInput:
      type: object
      required: [agencyId, message]
      properties:
        agencyId:
          type: integer
        itineraryId:
          type: integer
        message:
          type: string
    AgencyInquiryStatus:
      type: string
      enum: [pending, read]
    AgencyInquiry:
      type: object
      required: [id, agencyId, travelerId, message, status, createdAt]
      properties:
        id:
          type: integer
        agencyId:
          type: integer
        itineraryId:
          type: ["integer", "null"]
        travelerId:
          type: integer
        message:
          type: string
        status:
          $ref: "#/components/schemas/AgencyInquiryStatus"
        createdAt:
          type: string
...
    AgencyInquiryReceived:
      type: object
      required: [id, travelerName, travelerEmail, message, status, createdAt]
      properties:
        id:
          type: integer
        itineraryId:
          type: ["integer", "null"]
        itineraryName:
          type: ["string", "null"]
        travelerName:
          type: string
        travelerEmail:
          type: string
        message:
          type: string
        status:
          $ref: "#/components/schemas/AgencyInquiryStatus"
        createdAt:
          type: string
...
    Hotel:
      type: object
      required: [id, agencyId, name, city, country, active, createdAt]
      properties:
        id:
          type: integer
        agencyId:
          type: integer
        name:
          type: string
        city:
          type: string
        country:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        type:
          type: ["string", "null"]
        stars:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
    HotelInput:
      type: object
      required: [name, city, country]
      properties:
        name:
          type: string
        city:
          type: string
        country:
          type: string
        address:
          type: string
        phone:
          type: string
        website:
          type: string
        type:
          type: string
        stars:
          type: integer
        description:
          type: string
    HotelUpdate:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        country:
          type: string
        address:
          type: string
        phone:
          type: string
        website:
          type: string
        type:
          type: string
        stars:
          type: integer
        description:
          type: string
        active:
          type: boolean

    # ─── SEGMENT / DAY HOTEL ──────────────────────────────────────────────────
...
    ChecklistTemplate:
      type: object
      required: [id, agencyId, title, active, createdAt, updatedAt]
      properties:
        id:
          type: integer
        agencyId:
          type: integer
        title:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    ChecklistTemplateInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        active:
          type: boolean
    ChecklistTemplateUpdate:
      type: object
      properties:
        title:
          type: string
        active:
          type: boolean
...